+++ src/app/core/services/github.service.ts
@@ -26,6 +26,7 @@
 
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
+
 import { Observable } from 'rxjs/Observable';
 
 export class GithubUser {